Erring on the side of caution (for now). Stripped datapreview hack.
We were using a blacklist of filetypes which wouldn't be previewed; all
other items would be put in an iframe. Any remote binary file not on the
blacklist (BZ2, EXE...) would be automatically downloaded. Until we can
check the remote mimetype and be certain it is text, it's not safe to
put the file in an iframe.