Fix activity streams when no user logged in
Still get the user's ID from the revision, but fall back on 'not logged
in' if there is no user object in the revision. Put this logic in
lib.activity.

Add user_id param to lib.activity.activity_stream_item() and
model.package.Package.activity_stream_item().

Add a debug message when we skip an SQLAlchemy commit because it has no
_object_cache or no revision.

Add test cases for adding and updating packages and resources when not logged in.