We can now create custom groups using the registered group_type from the plugin and /type/new. Also implemented named routes for group controller.
When registered the custom group form registers new routes for the types defined in group_types()
so that they can be accessed via a top-level url. This uses named routes for the groups and the
custom group types, although no checking is done yet to make sure it doesn't overwrite a top level
object.