[#1494][#1496] Save pkg and res in activity tables
When recording an activity or activity detail for a new or changed
package or resource, save a copy of the package or resource (dictized
and dumped as json) to the 'data' column of the activity or
activity_detail table.

In the activity stream templates get package and resource titles etc.
from this data, instead of querying the controller functions for them.

This means that if a package or resource name changes, old activity
events will still show the name from the time of the activity, if a
package or resource gets deleted old activity events will still have
access to the data, etc.

It also reduces extra database accesses when rendering activity streams.

And it means that this data about the package or resource that the
activity relates to is available in the json from the api call.