[#1636] Temporary workaround for activity streams crash when purging objects
The problem is that whenever a package gets updated the activity streams
session tries to dictize that package in order to write a copy of the
dict into the activity streams table, but if the package is being purged
then package_dictize() raises an exception when it can't find a current
revision for the package.

This fix just skips making an activity stream item for a package that is
being purged, which is probably the right thing to do, but perhaps we
should also write code to purge any previous activity stream items
related to the purged package.