[#1547][plugins,search] Add dataset search plugin interface
The IPackageController now also offers the `before_search` and
`after_search` methods so extensions can modify the search parameters
and the results coming from the standard Solr search.

Parameters starting with `ext_` will be added to an `extras` property
which will have no effect in the default search but will be forwarded
to the extensions as part of the search parameters.

Extensions willing to modify the standard search form can do so
implementing IGenshiStreamFilter and injecting their extra fields
in the `dataset-search-ext` div.