[master][#1547] Abort search flag
If extensions add an `abort_search` = True pair to the search_params
dict returned by the `before_search` plugin hook, the actual query
to Solr will be skiped and 0 results will be returned.

Useful e.g. if the extension has performed a previous query that returned
no records and the Solr one is no longer needed.