[#1666] Add tag activities to activity streams
Adds 'USER added/removed tag FOO to/from dataset BAR' activities to user
and dataset activity streams.

Currently whenever the activity streams session extension receives a
'changed PackageTag' event a 'removed tag' activity is emitted, assuming
that the change to the package tag was to set it's state to deleted. But
eventually we need proper detection of deletes versus changes, for other
objects (e.g. resources) as well as for tags.

Tests for this still need to be written.

'Added tag' and 'removed tag' activities show up as 'changed package'
activities with one 'added tag' or 'removed tag' activity detail. These
get rendered simply as 'USER added/removed the tag TAG to/from the
dataset DATASET' because when a 'changed package' activity has only one
activity detail that detail is rendered in place of the top-level
activity. (This is the same way that 'added resource' and 'updated
resource' activities work.)

If you make several changes to a dataset at once (e.g. adding and/or
removing multiple tags, or adding and/or removing tags and making other
changes) it shows up as a 'changed package' activity with multiple
activity details, and is rendered as a 'changed package' activity.

Details:

Add activity_stream_detail() function to PackageTag model class. Tag
events are generated by the activity streams session extension, not by
the logic functions.

Add added and removed tags rendering in ckan/logic/action/get.py, and
templates to ckan/templates/activity_streams/

The new tag activities already appear in the user and dataset's activity
streams.