Fix handling of deleted group activity streams
They now show up as 'deleted group' activities instead of 'changed
group'.

The implemenetation is a bit of a hack.

The group_update() logic function gets called when you delete a group
through the web interface, but there is also a group_delete() logic
function, when is this used?

Also add a test/