[#1639] Add some basic CSS to activity streams
Just so they look better than being completely unstyled, trying to make
them a little better to scan with the eyes: emphasise the actor, object
and (to a lesser extend) the verb, demphasise the date, and add some
vertical space in-between activities.

Also tidied up the HTML, the CSS classes, and the Genshi macro a little.