[release-1.6.1][solr] Added a new field, titleString, to the solr schema.
Since the 'title' field is of type 'text', it means we can't sort on it.  So I've added a new
solr field that has a copy of the 'title' field, but of type 'string', meaning it can be
sorted on.

There are other 'text' fields, but I think the 'title' is likely the only one we'd want to
sort on.

This change is backwards compatible, hence not creating a new version of the schema