[2332] Fix width of navigation on resources page
This was actually an issue with the rest of the page being wider than
the all the others making the navigation look narrower. I've reduced
the width to 940px, this was originally set to 950px by Rufus as a fix
to #1396 but change does not seem to affect the page.