[#2304] Small user controller refactoring
Move initialisation of a few more template variables into
_setup_template_variables(), and change read(), edit(), and followers()
to use it. This removes some code duplication and fixes issues with the
followers count and follow button not being initialisd on all user
controller pages.

Change new() to _not_ use _setup_template_variables() as it only needs
c.is_sysadmin and not the rest.

Also fix templates/user/layout.html so that the Followers tab appears on
both your own user page (when logged in) and on other user's pages.