2322 Added some simple convenience urls
Added changes to routes for cosmetically improved urls with a
redirect for the API call, a new package controller function for
redirecting to the actual resource url if it exists.

The /dataset/id/resource/id/viewer url uses the existing embed url
with an improved default that more closely matches the screen size
although doesn't change the embed code to use % instead of px.