[#3028] Add 'num_resources' to package dicts
Add 'num_resources' to the dicts returned by package_show().

This is convenient for some javascript that has to show package
previews.

This fixes an issue where a followed dataset's number of resources would
not show up when that dataset was previewed in the sidebar on the user
dashboard page.