[#3028] Fix user_activity_list to accept name or id
user_activity_list should accept either a user name or user id for the
'id' arg, but was only accepting IDs. Fix it to accept either, and to
raise NotFound if the name or id does not exist.

This fixes user activity streams not showing in the user dashboard, when
the dashboard was filtered to a user.