Add about_template() to DefaultGroupForm
It looks as if commit e51e9146508fd6da27be6d662d603db2295d6c03 changed
GroupController to call a new about_template() method of group plugins,
but about_template() was not added to the default group plugin (class
DefaultGroupForm), so it would crash on trying to call it.

This was causing a crash when trying to add a related item to a dataset
on publicdata.eu, for example.