Run DataStore tests in Travis-CI
I had to do a workaround so "paster datastore set-permissions" would work. The
problem is that the code calls psql passing the -W option. This option asks for
the database's password, even if you don't need it. So I simply removed it.

We should look into a better solution for this.