[#255] Edit FileStore API docs
Move the FileStore API docs into its own file (filestore-api.rst) same
as the Action API and DataStore API, and move it to the CKAN API section
in the table of contents.

Also renamed it Storage API -> FileStore API (since the feature is
called the FileStore), and added a link to the ckanclient code that
shows how to use it (since the documentation is hopeless).

On the FileStore page, where the FileStore API docs used to be put a
link to the new FileStore API page.

Edited the FileStore feature docs: rename Storage Web Interface ->
FileStore Web Interface (the feature seems to change names, from
FileStore to just Storage, half-way down the page).