db_to_form_package_schema() don't strip keys from resource dicts
Add several keys to the 'resources' subschema in
db_to_form_package_schema() so that validation doesn't strip the keys
from the resources dicts. This fixes a test that was failing if run with
an IDatasetForm plugin active, and may also fix some template crashes
when an IDatasetForm plugin is active if people are accessing these
resource keys in their templates.

Fixes #415.