Revert "[#417] Don't pass api version to form_to_db_schema_options()"
This reverts commit e023d8a5f50baf3931e2610eaa433b5653227154.

e023d8a5 was the wrong fix and broke a bunch of tests, package_update()
was only passing 'api': True if it was in fact called via the api, which
is what it should do.

The real bug is in the package controller's _save_edit() method, which
puts the api version in the context even though it's a web ui method not
an api method. I'll fix that in the next commit...