[#716] Allow to pass the schema as part of the context
In some cases extensions need to tweak the schema used on package
creation or update. For instance harvesters may want to create a package
with a certain id, or relax the default tags schema. There used to be an
option via `form_to_db_schema_options` to provide a custom schema via
the context, but this was removed on the last IDatasetForm refactor.

This adds back the option to pass the schema as part of the context.