[#519] Update source install docs for 2.0
Update the source install and related docs (datastore setup, solr setup,
testing) and files (deployment.ini_tmp, test-core.ini) for CKAN 2.0.

The idea is to document the setup process that OKFN is currently using
on its servers, so that the deployment docs can become the docs for the
OKFN dev team and we can dogfood our own docs, instead of having a bunch
of wiki pages, READMEs, etc. that we follow and ignoring the "official"
docs. Hopefully those can all be deleted or cut down massively once this
is done.

So both the docs now describe the layout we use on our servers, where
there's space for multiple CKAN sites on the same server, with
virtualenvs in /usr/lib/ckan/ and config files in /etc/ckan/.

- Changed default virtualenv location from `~/pyenv` to `/usr/lib/ckan/default`

- Changed default config files dir to `/etc/ckan/default`

- Added tip for developers in source install docsS symlink `/usr/lib/ckan` and
  `/etc/ckan` to directories in your homedir to install ckan in your homedir

- Changed default database user from `ckanuser` to `ckan_default`

- Changed default database from `ckan_dev` to `ckan_default`

- Changed default datastore database from `datastore` to `datastore_default`

- Changed default datastore database user from `readonlyuser` to `datastore_default`

- Changed default test databases from `ckan_test` and `ckan_test_datastore` to
  `ckan_test` and `datastore_test`

- Changed default `ckan.site_id` from `ckan.net` to `default`

- Disabled file log handler, rely on Apache's log files only (as we do on our
  servers)

- Changed the docs to use restructured text substitutions for things like the
  paths to the virtualenv and config files, database names, etc. instead of
  typing them out each time

- Lots of minor formatting fixes and editing

So now we have:

    virtualenv:        /usr/lib/ckan/default
    development.ini:   /etc/ckan/default/development.ini
    production.ini:    /etc/ckan/default/production.ini
    ckan.site_id:      default
    Database user:     ckan_default
    Database:          ckan_default
    Datastore user:    datastore_default
    Datastore db:      datastore_default
    Test database:     ckan_test
    Test datastore db: datastore_test